+++ source/benoit/benoit.rs
@@ -35,8 +35,8 @@ pub mod video;
 
 pub const VERSION: (u32, u32, u32) = (
 	0x2, // Major
-	0x5, // Minor
-	0x1, // Patch
+	0x6, // Minor
+	0x0, // Patch
 );
 
 pub const PRECISION: u32 = 0x80;