+++ source/benoit/benoit/script/dump_frame.rs
@@ -34,7 +34,7 @@ use rug::Float;
 use std::time::Instant;
 
 impl Script {
-	pub(super) fn dump(
+	pub(super) fn dump_frame(
 		dump_path:      &str,
 		name:           &str,
 		image:          &mut Image,
@@ -48,7 +48,8 @@ impl Script {
 		colour_range:   f32,
 		image_format:   ImageFormat,
 	) {
-		eprint!("\"{name}\" (2^{:.9}x)...", zoom.to_f64().log2());
+		eprint!("\"{name}\" (2^{:.9}x): rendering...", zoom.to_f64().log2());
+		Script::set_title("rendering");
 
 		let time_start = Instant::now();
 
@@ -62,15 +63,19 @@ impl Script {
 
 		let render_time = time_start.elapsed();
 		eprint!(" {:.3}ms, colouring...", render_time.as_micros() as f32 / 1000.0);
+		Script::set_title("colouring");
 
 		image.colour(&render, palette, max_iter_count, colour_range);
 
 		let colour_time = time_start.elapsed() - render_time;
-		eprint!(" {:.3}ms...", colour_time.as_micros() as f32 / 1000.0);
+		eprint!(" {:.3}ms, dumping...", colour_time.as_micros() as f32 / 1000.0);
+		Script::set_title("dumping");
 
 		let path = format!("{dump_path}/{name}");
-
 		image.dump(path.as_str(), image_format);
-		eprintln!(" done");
+
+		let dump_time = time_start.elapsed() - colour_time - render_time;
+		eprintln!(" {:.3}ms - \u{1B}[1m\u{1B}[92mdone!\u{1B}[0m", dump_time.as_micros() as f32 / 1000.0);
+		Script::set_title("done");
 	}
 }